% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barChart.r
\name{barChart}
\alias{barChart}
\title{Produces bar charts.}
\usage{
barChart()
}
\value{
Creates bar charts.
}
\description{
Makes bar graphs showing the number of genes under peaks at various upstream extension levels.
}
\examples{
\dontrun{generate(rat, 1000, 3000, 100)}
\dontrun{barChart()}

}

