% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffGO.R
\name{diffGO}
\alias{diffGO}
\title{Finds differential gene ontologies}
\usage{
diffGO(organism, start, end, GOcategory, GOspecies)
}
\arguments{
\item{organism}{Object name assigned from readGFF() command.}

\item{start}{Lower bound of upstream extension.}

\item{end}{Upper bound of upstream extension.}

\item{GOcategory}{Either BP, CC, or MF.}

\item{GOspecies}{Optional, can be used to limit the gene hits from given species. You can use "common names" for nine common species (human, mouse, rat, fruitfly, nematode, zebrafish, thale-cress, frog and pig), this common name is allowed to be identical to the organism object parameter. For all other species, you can provide their taxonomy ids.  Multiple species can be passed using comma as a separator. Passing "all" will query against all available species. Default: human,mouse,rat.}
}
\value{
A character vector of gene ontology (BP, CC, or MF) terms for unique genes located under peaks between two upstream extension levels.
}
\description{
Determines gene ontology terms for each category (biological process (BP), cellular compartment (CC), molecular function (MF)) of genes-under-peaks that are unique between two different upstream extension levels.
}
\examples{
rat <- readGFF("ftp://ftp.ensembl.org/pub/release-84/gtf/rattus_norvegicus/Rattus_norvegicus.Rnor_6.0.84.gtf.gz")
fpath <- system.file("extdata", "somepeaksfile.txt", package="geneXtendeR")
peaksInput(fpath)
diffGO(rat, 2000, 3000, BP, rat)

}
