% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneXtender.r
\name{.geneXtender}
\alias{.geneXtender}
\title{Function for formatting data to be called by C function.}
\usage{
.geneXtender(upstream, organism, ret = FALSE)
}
\arguments{
\item{upstream}{Extension upstream of each gene.}

\item{organism}{Object name assigned from readGFF() command.}

\item{ret}{Logical that designates if the output is written in R or only as a .bed file}
}
\value{
The gene coordinates extended by the designated upstream. In .bed file format.

A data.table formatted version of the data depending on the ret logical.
}
\description{
Formats genome-data based on upstream and organism for .C caller function. Not to be run by user.
}
\note{
Not to be run by the user
}
\keyword{internal}
