% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.r
\docType{data}
\name{mouse}
\alias{mouse}
\title{Gene transfer format (GTF) file for mouse (Mus_musculus.GRCm38.93)}
\format{A data frame with 1793461 rows and 26 variables corresponding to the entries of a GTF file.  Column names are standardized and can be found here: http://www.ensembl.org/info/website/upload/gff.html.}
\usage{
data(mouse)
}
\value{
Demonstrates a mouse GTF file downloaded from: ftp://ftp.ensembl.org/pub/release-93/gtf/mus_musculus/Mus_musculus.GRCm38.93.chr.gtf.gz
}
\description{
A dataset downloaded from Ensembl that contains the entries of a GTF file for Mus musculus.
}
\examples{
head(mouse)
tail(mouse)

}
\keyword{datasets}
