% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat.r
\docType{data}
\name{rat}
\alias{rat}
\title{Gene transfer format (GTF) file for rat (Rattus_norvegicus.Rnor_6.0.84)}
\format{A data frame with 748514 rows and 28 variables corresponding to the entries of a GTF file.  Column names are standardized and can be found here: http://www.ensembl.org/info/website/upload/gff.html.}
\usage{
data(rat)
}
\value{
Demonstrates a rat GTF file downloaded from: ftp://ftp.ensembl.org/pub/release-84/gtf/rattus_norvegicus/Rattus_norvegicus.Rnor_6.0.84.gtf.gz.
}
\description{
A dataset downloaded from Ensembl that contains the entries of a GTF file for Rattus norvegicus.
}
\examples{
head(rat)
tail(rat)

}
\keyword{datasets}

