% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_input_docs.r
\docType{data}
\name{sample_peaks_input}
\alias{sample_peaks_input}
\title{Peaks list ordered by chromosome number}
\format{A data frame with 9799 rows and 3 variables:
\describe{
\item{V1}{Chromosome number}
\item{V2}{Peak start position [in units of base pairs]}
\item{V3}{Peak end position [in units of base pairs]}
}}
\usage{
data(sample_peaks_input)
}
\description{
A dataset containing the start and stop positions
of ChIP-seq peaks along a genome
}
\examples{
head(sample_peaks_input)
tail(sample_peaks_input)

}
\keyword{datasets}

