% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_input_docs.r
\docType{data}
\name{samplepeaksinput}
\alias{samplepeaksinput}
\title{Peaks list ordered by chromosome number and start position}
\format{A data frame with 25088 rows and 3 variables:
\describe{
\item{V1}{Chromosome number}
\item{V2}{Peak start position [in units of base pairs]}
\item{V3}{Peak end position [in units of base pairs]}
}}
\usage{
data(samplepeaksinput)
}
\value{
Demonstrates a sample peaks file used as input.  No column headers are present to ensure proper sorting in subsequent analysis steps.
}
\description{
A dataset containing the chromosome number, start and stop positions
of ChIP-seq peaks along a genome
}
\examples{
head(samplepeaksinput)
tail(samplepeaksinput)

}
\keyword{datasets}

