% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.r
\name{species}
\alias{species}
\title{Show available species}
\usage{
species()
}
\value{
Species available for geneXtendeR analysis.
}
\description{
Show available species from the gtf package for geneXtendeR analysis.
}
\examples{
species()

}

