% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{annotate}
\alias{annotate}
\title{Annotate peaks file.}
\usage{
annotate(organism, extension)
}
\arguments{
\item{organism}{Object name assigned from readGFF() command.}

\item{extension}{Desired upstream extension.}
}
\value{
A file of peaks annotated with gene ID, gene name, and gene-to-peak genomic distance (in bp).
}
\description{
Annotate peaks file with gene information based on optimally chosen geneXtendeR upstream extension file.
}
\examples{
rat <- readGFF("ftp://ftp.ensembl.org/pub/release-84/gtf/rattus_norvegicus/Rattus_norvegicus.Rnor_6.0.84.gtf.gz")
peaksInput("somepeaksfile.txt")
annotate(rat, 2500)

}

