% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barChart.r
\name{barChart}
\alias{barChart}
\title{Produces bar charts.}
\usage{
barChart(organism, start, end, by)
}
\value{
Creates bar charts.
}
\description{
Makes bar graphs showing the number of genes under peaks at various upstream extension levels.
}
\examples{
\dontrun{mouse <- readGFF("ftp://ftp.ensembl.org/pub/release-84/gtf/mus_musculus/Mus_musculus.GRCm38.84.gtf.gz")}
\dontrun{peaksInput("somepeaksfile.txt")}
\dontrun{barChart(mouse, 1000, 3000, 100)}

}

