% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{generate}
\alias{generate}
\title{Generate geneXtendeR files.}
\usage{
generate(organism, start, end, by)
}
\arguments{
\item{organism}{Object name assigned from readGFF() command.}

\item{start}{Lower bound of upstream extension.}

\item{end}{Upper bound of upstream extension.}

\item{by}{Interval between consecutive extensions.}
}
\value{
Generates exactly ([(start-end)/by] + 1) modified gtf files and loads in global variables for subsequent analysis.
}
\description{
Creates organism-specific customized gene transfer format files at user-designated extensions.  Only upstream extensions can be specified.  Downstream extensions are always set as default to 500 bp (to capture any downstream gene elements).
}
\examples{
\dontrun{human <- readGFF("ftp://ftp.ensembl.org/pub/release-84/gtf/homo_sapiens/Homo_sapiens.GRCh38.84.gtf.gz")}
\dontrun{generate(human, 2000, 10000, 500)}

}

