% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linePlot.r
\name{linePlot}
\alias{linePlot}
\title{Produces line plots.}
\usage{
linePlot()
}
\value{
Creates differential line plots.
}
\description{
Makes differential line plots showing the differences in the number of genes under peaks at consecutive upstream extension levels.
}
\examples{
\dontrun{generate(rat, 1000, 3000, 100)}
\dontrun{linePlot()}  #Explained in vignette

}

