% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksInput.r
\name{peaksInput}
\alias{peaksInput}
\title{Formats a peaks input file.}
\usage{
peaksInput(filename)
}
\arguments{
\item{filename}{Name of file containing peaks.  This tab-delimited input file may only contain 3 columns: chromosome number, peak start, and peak end.  See ?samplepeaksinput for more details.}
}
\value{
Returns a formatted file to be used with barChart(), linePlot(), and other downstream commands.
}
\description{
Takes your tab-delimited 3-column input file (see Arguments section below) consisting of peaks called from a peak caller (e.g., MACS2 or SICER) and sorts the file by chromosome and start position to create a formatted file for downstream geneXtendeR analysis.  For sorting purposes, the X chromosome is designated by the integer 100, the Y chromosome by the integer 200, and the mitochondrial chromosome by the integer 300.
}
\examples{
?samplepeaksinput  #Documentation of some exemplary sample input
data(samplepeaksinput)
head(samplepeaksinput)
tail(samplepeaksinput)
peaksInput("somepeaksfile.txt")

}

